Split storage service workflows
